+++ Changelog.md
@@ -5,6 +5,7 @@
 - Dropped Laravel 9 support
 - Dropped PHP 8.0 & 8.1 support
 - Added PHP 8.3 support
+- Added the `Ownable` interface to be used by Models
 - Added the `HasRoles::havingPermission()` query scope that accept a single `string|Permission` parameter
 - Added the `HasRoles::havingRole()` query scope that accept a single `string|Role` parameter
 - BC: The 'Permission::findByName()' method no longer throws an exception when a permission was not found, but returns NULL instead